+++ coredb-operator/src/lib.rs
@@ -16,6 +16,8 @@ mod postgres_exporter_role;
 mod psql;
 mod secret;
 mod service;
+mod servicemonitor;
+mod servicemonitor_crd;
 mod statefulset;
 use thiserror::Error;
