--- coredb-operator/src/servicemonitors.rs
@@ -1,394 +1,0 @@
-// WARNING: generated by kopium - manual changes will be overwritten
-// kopium command: kopium servicemonitors.monitoring.coreos.com
-// kopium version: 0.14.0
-
-use kube::CustomResource;
-use serde::{Serialize, Deserialize};
-use std::collections::BTreeMap;
-use k8s_openapi::apimachinery::pkg::util::intstr::IntOrString;
-
-#[derive(CustomResource, Serialize, Deserialize, Clone, Debug)]
-#[kube(group = "monitoring.coreos.com", version = "v1", kind = "ServiceMonitor", plural = "servicemonitors")]
-#[kube(namespaced)]
-#[kube(schema = "disabled")]
-pub struct ServiceMonitorSpec {
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "attachMetadata")]
-    pub attach_metadata: Option<ServiceMonitorAttachMetadata>,
-    pub endpoints: Vec<ServiceMonitorEndpoints>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "jobLabel")]
-    pub job_label: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "labelLimit")]
-    pub label_limit: Option<i64>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "labelNameLengthLimit")]
-    pub label_name_length_limit: Option<i64>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "labelValueLengthLimit")]
-    pub label_value_length_limit: Option<i64>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "namespaceSelector")]
-    pub namespace_selector: Option<ServiceMonitorNamespaceSelector>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "podTargetLabels")]
-    pub pod_target_labels: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "sampleLimit")]
-    pub sample_limit: Option<i64>,
-    pub selector: ServiceMonitorSelector,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "targetLabels")]
-    pub target_labels: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "targetLimit")]
-    pub target_limit: Option<i64>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorAttachMetadata {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub node: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpoints {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub authorization: Option<ServiceMonitorEndpointsAuthorization>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "basicAuth")]
-    pub basic_auth: Option<ServiceMonitorEndpointsBasicAuth>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "bearerTokenFile")]
-    pub bearer_token_file: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "bearerTokenSecret")]
-    pub bearer_token_secret: Option<ServiceMonitorEndpointsBearerTokenSecret>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "enableHttp2")]
-    pub enable_http2: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "filterRunning")]
-    pub filter_running: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "followRedirects")]
-    pub follow_redirects: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "honorLabels")]
-    pub honor_labels: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "honorTimestamps")]
-    pub honor_timestamps: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub interval: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "metricRelabelings")]
-    pub metric_relabelings: Option<Vec<ServiceMonitorEndpointsMetricRelabelings>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub oauth2: Option<ServiceMonitorEndpointsOauth2>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub params: Option<BTreeMap<String, String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub port: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "proxyUrl")]
-    pub proxy_url: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub relabelings: Option<Vec<ServiceMonitorEndpointsRelabelings>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scheme: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "scrapeTimeout")]
-    pub scrape_timeout: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "targetPort")]
-    pub target_port: Option<IntOrString>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "tlsConfig")]
-    pub tls_config: Option<ServiceMonitorEndpointsTlsConfig>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsAuthorization {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub credentials: Option<ServiceMonitorEndpointsAuthorizationCredentials>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "type")]
-    pub r#type: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsAuthorizationCredentials {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsBasicAuth {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub password: Option<ServiceMonitorEndpointsBasicAuthPassword>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub username: Option<ServiceMonitorEndpointsBasicAuthUsername>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsBasicAuthPassword {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsBasicAuthUsername {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsBearerTokenSecret {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsMetricRelabelings {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub action: Option<ServiceMonitorEndpointsMetricRelabelingsAction>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub modulus: Option<i64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub regex: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub replacement: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub separator: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "sourceLabels")]
-    pub source_labels: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "targetLabel")]
-    pub target_label: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum ServiceMonitorEndpointsMetricRelabelingsAction {
-    #[serde(rename = "replace")]
-    Replace,
-    Replace,
-    #[serde(rename = "keep")]
-    Keep,
-    Keep,
-    #[serde(rename = "drop")]
-    Drop,
-    Drop,
-    #[serde(rename = "hashmod")]
-    Hashmod,
-    HashMod,
-    #[serde(rename = "labelmap")]
-    Labelmap,
-    LabelMap,
-    #[serde(rename = "labeldrop")]
-    Labeldrop,
-    LabelDrop,
-    #[serde(rename = "labelkeep")]
-    Labelkeep,
-    LabelKeep,
-    #[serde(rename = "lowercase")]
-    Lowercase,
-    Lowercase,
-    #[serde(rename = "uppercase")]
-    Uppercase,
-    Uppercase,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsOauth2 {
-    #[serde(rename = "clientId")]
-    pub client_id: ServiceMonitorEndpointsOauth2ClientId,
-    #[serde(rename = "clientSecret")]
-    pub client_secret: ServiceMonitorEndpointsOauth2ClientSecret,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "endpointParams")]
-    pub endpoint_params: Option<BTreeMap<String, String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scopes: Option<Vec<String>>,
-    #[serde(rename = "tokenUrl")]
-    pub token_url: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsOauth2ClientId {
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "configMap")]
-    pub config_map: Option<ServiceMonitorEndpointsOauth2ClientIdConfigMap>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub secret: Option<ServiceMonitorEndpointsOauth2ClientIdSecret>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsOauth2ClientIdConfigMap {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsOauth2ClientIdSecret {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsOauth2ClientSecret {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsRelabelings {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub action: Option<ServiceMonitorEndpointsRelabelingsAction>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub modulus: Option<i64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub regex: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub replacement: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub separator: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "sourceLabels")]
-    pub source_labels: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "targetLabel")]
-    pub target_label: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum ServiceMonitorEndpointsRelabelingsAction {
-    #[serde(rename = "replace")]
-    Replace,
-    Replace,
-    #[serde(rename = "keep")]
-    Keep,
-    Keep,
-    #[serde(rename = "drop")]
-    Drop,
-    Drop,
-    #[serde(rename = "hashmod")]
-    Hashmod,
-    HashMod,
-    #[serde(rename = "labelmap")]
-    Labelmap,
-    LabelMap,
-    #[serde(rename = "labeldrop")]
-    Labeldrop,
-    LabelDrop,
-    #[serde(rename = "labelkeep")]
-    Labelkeep,
-    LabelKeep,
-    #[serde(rename = "lowercase")]
-    Lowercase,
-    Lowercase,
-    #[serde(rename = "uppercase")]
-    Uppercase,
-    Uppercase,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfig {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub ca: Option<ServiceMonitorEndpointsTlsConfigCa>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "caFile")]
-    pub ca_file: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cert: Option<ServiceMonitorEndpointsTlsConfigCert>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "certFile")]
-    pub cert_file: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "insecureSkipVerify")]
-    pub insecure_skip_verify: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "keyFile")]
-    pub key_file: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "keySecret")]
-    pub key_secret: Option<ServiceMonitorEndpointsTlsConfigKeySecret>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "serverName")]
-    pub server_name: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigCa {
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "configMap")]
-    pub config_map: Option<ServiceMonitorEndpointsTlsConfigCaConfigMap>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub secret: Option<ServiceMonitorEndpointsTlsConfigCaSecret>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigCaConfigMap {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigCaSecret {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigCert {
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "configMap")]
-    pub config_map: Option<ServiceMonitorEndpointsTlsConfigCertConfigMap>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub secret: Option<ServiceMonitorEndpointsTlsConfigCertSecret>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigCertConfigMap {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigCertSecret {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorEndpointsTlsConfigKeySecret {
-    pub key: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorNamespaceSelector {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub any: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "matchNames")]
-    pub match_names: Option<Vec<String>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorSelector {
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "matchExpressions")]
-    pub match_expressions: Option<Vec<ServiceMonitorSelectorMatchExpressions>>,
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "matchLabels")]
-    pub match_labels: Option<BTreeMap<String, String>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ServiceMonitorSelectorMatchExpressions {
-    pub key: String,
-    pub operator: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub values: Option<Vec<String>>,
-}
-
